(PC-27682)[PRO] refactor: merge screen and form folders into pages
