+++ pro/src/pages/Offerers/Offerer/VenueV1/VenueEdition/CollectiveDataEdition/CollectiveDataEdition.tsx
@@ -9,12 +9,12 @@ import {
 } from 'apiClient/v1'
 import Callout from 'components/Callout/Callout'
 import MandatoryInfo from 'components/FormLayout/FormLayoutMandatoryInfo'
-import { CollectiveDmsTimeline } from 'components/VenueCreationForm/CollectiveDmsTimeline/CollectiveDmsTimeline'
 import { getEducationalDomainsAdapter } from 'core/OfferEducational'
 import { GET_DATA_ERROR_MESSAGE } from 'core/shared'
 import { SelectOption } from 'custom_types/form'
 import useNotification from 'hooks/useNotification'
 import { PartnerPageCollectiveSection } from 'pages/Home/Offerers/PartnerPageCollectiveSection'
+import { CollectiveDmsTimeline } from 'pages/VenueCreation/CollectiveDmsTimeline/CollectiveDmsTimeline'
 import Spinner from 'ui-kit/Spinner/Spinner'
 import { getLastCollectiveDmsApplication } from 'utils/getLastCollectiveDmsApplication'
 import { sendSentryCustomError } from 'utils/sendSentryCustomError'
