+++ pro/src/pages/Offerers/Offerer/VenueV1/fields/ReimbursementFields/ReimbursementFields.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react'
 
 import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import FormLayout from 'components/FormLayout'
-import ReimbursementPoint from 'components/VenueCreationForm/ReimbursementPoint/ReimbursementPoint'
+import ReimbursementPoint from 'pages/VenueCreation/ReimbursementPoint/ReimbursementPoint'
 import useActiveFeature from 'hooks/useActiveFeature'
 import { Banner } from 'ui-kit'
