+++ pro/src/pages/VenueCreation/Informations/SiretOrCommentFields/SiretOrCommentFields.tsx
@@ -4,9 +4,9 @@ import React, { useEffect, useState } from 'react'
 import { getAdressDataAdapter } from 'components/Address/adapter'
 import { handleAddressSelect } from 'components/Address/Address'
 import FormLayout from 'components/FormLayout'
-import { VenueCreationFormValues } from 'components/VenueCreationForm/types'
 import getSiretData from 'core/Venue/adapters/getSiretDataAdapter'
 import { humanizeSiret, unhumanizeSiret } from 'core/Venue/utils'
+import { VenueCreationFormValues } from 'pages/VenueCreation/types'
 import { InfoBox, TextArea, TextInput } from 'ui-kit'
 import Toggle from 'ui-kit/Toggle'
