+++ pro/src/pages/VenueCreation/VenueCreationFormScreen.tsx
@@ -5,10 +5,6 @@ import { useLocation, useNavigate } from 'react-router-dom'
 import { api } from 'apiClient/api'
 import { isErrorAPIError, serializeApiErrors } from 'apiClient/helpers'
 import { GetOffererResponseModel, VenueProviderResponse } from 'apiClient/v1'
-import { generateSiretValidationSchema } from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
-import { VenueCreationFormValues } from 'components/VenueCreationForm/types'
-import { validationSchema } from 'components/VenueCreationForm/validationSchema'
-import { VenueCreationForm } from 'components/VenueCreationForm/VenueCreationForm'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
 import { Providers } from 'core/Venue/types'
@@ -16,11 +12,15 @@ import { SelectOption } from 'custom_types/form'
 import useAnalytics from 'hooks/useAnalytics'
 import useCurrentUser from 'hooks/useCurrentUser'
 import useNotification from 'hooks/useNotification'
+import { generateSiretValidationSchema } from 'pages/VenueCreation/Informations/SiretOrCommentFields'
+import { VenueCreationFormValues } from 'pages/VenueCreation/types'
+import { validationSchema } from 'pages/VenueCreation/validationSchema'
+import { VenueCreationForm } from 'pages/VenueCreation/VenueCreationForm'
 import { Title } from 'ui-kit'
 
 import { serializePostVenueBodyModel } from './serializers'
-import { venueSubmitRedirectUrl } from './utils/venueSubmitRedirectUrl'
 import style from './VenueCreationFormScreen.module.scss'
+import { venueSubmitRedirectUrl } from './venueSubmitRedirectUrl'
 
 interface VenueEditionProps {
   initialValues: VenueCreationFormValues
