+++ pro/src/pages/VenueEdition/VenueEditionFormScreen.tsx
@@ -7,10 +7,6 @@ import { isErrorAPIError, serializeApiErrors } from 'apiClient/helpers'
 import { GetOffererResponseModel, GetVenueResponseModel } from 'apiClient/v1'
 import Callout from 'components/Callout/Callout'
 import ConfirmDialog from 'components/Dialog/ConfirmDialog'
-import { generateSiretValidationSchema } from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
-import { VenueEditionFormValues } from 'components/VenueEditionForm/types'
-import { validationSchema } from 'components/VenueEditionForm/validationSchema'
-import { VenueEditionForm } from 'components/VenueEditionForm/VenueEditionForm'
 import { Events } from 'core/FirebaseEvents/constants'
 import { PATCH_SUCCESS_MESSAGE } from 'core/shared'
 import { SelectOption } from 'custom_types/form'
@@ -19,8 +15,12 @@ import useCurrentUser from 'hooks/useCurrentUser'
 import useNotification from 'hooks/useNotification'
 import strokeMailIcon from 'icons/stroke-mail.svg'
 import { PartnerPageIndividualSection } from 'pages/Home/Offerers/PartnerPageIndividualSection'
+import { generateSiretValidationSchema } from 'pages/VenueCreation/Informations/SiretOrCommentFields'
 
 import { serializeEditVenueBodyModel } from './serializers'
+import { VenueEditionFormValues } from './types'
+import { validationSchema } from './validationSchema'
+import { VenueEditionForm } from './VenueEditionForm'
 import styles from './VenueEditionFormScreen.module.scss'
 
 interface VenueEditionProps {
