+++ pro/src/pages/VenueEdition/validationSchema.ts
