+++ pro/src/pages/VenueSettings/VenueSettingsForm.tsx
@@ -8,11 +8,11 @@ import {
 } from 'apiClient/v1'
 import { AddressSelect } from 'components/Address'
 import FormLayout from 'components/FormLayout'
-import BankAccountInfos from 'components/VenueCreationForm/BankAccountInfos/BankAccountInfos'
-import SiretOrCommentFields from 'components/VenueCreationForm/Informations/SiretOrCommentFields'
-import { OffersSynchronization } from 'components/VenueCreationForm/OffersSynchronization'
-import { VenueFormActionBar } from 'components/VenueCreationForm/VenueFormActionBar'
-import { WithdrawalDetails } from 'components/VenueCreationForm/WithdrawalDetails'
+import BankAccountInfos from 'pages/VenueCreation/BankAccountInfos/BankAccountInfos'
+import SiretOrCommentFields from 'pages/VenueCreation/Informations/SiretOrCommentFields'
+import { OffersSynchronization } from 'pages/VenueCreation/OffersSynchronization'
+import { VenueFormActionBar } from 'pages/VenueCreation/VenueFormActionBar'
+import { WithdrawalDetails } from 'pages/VenueCreation/WithdrawalDetails'
 import { Providers } from 'core/Venue/types'
 import { SelectOption } from 'custom_types/form'
 import { useScrollToFirstErrorAfterSubmit } from 'hooks'
